requirements.txt
azure-functions==1.0.8
PyYAML==5.1
cassandra-driver==3.21.0
pandas==0.24.2
numpy==1.16.4
azure-storage-blob==12.2.0
azure-storage==0.36.0
